# Conflicts:
#	app/src/main/java/com/tosslab/jandi/app/network/dagger/ApiClientComponent.java
#	app/src/main/java/com/tosslab/jandi/app/push/PushInterfaceActivity.java
#	app/src/main/java/com/tosslab/jandi/app/ui/account/AccountHomeActivity.java
#	app/src/main/java/com/tosslab/jandi/app/ui/account/presenter/AccountHomePresenterImpl.java
#	app/src/main/java/com/tosslab/jandi/app/ui/maintab/MainTabActivity.java
#	app/src/main/res-crawrer/values-ko/strings.xml

+++ app/src/main/java/com/tosslab/jandi/app/events/entities/MemberStarredEvent.java
@@ -5,13 +5,13 @@ package com.tosslab.jandi.app.events.entities;
  */
 public class MemberStarredEvent {
 
-    private final int id;
+    private final long id;
 
-    public MemberStarredEvent(int id) {
+    public MemberStarredEvent(long id) {
         this.id = id;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 }
